Infer PHP for older WordPress URLs
